Feature/add backend testing suite

+++ backend/controllers/user.controller.ts
@@ -95,30 +95,42 @@ const loginUser = asyncHandler(async (req: Request, res: Response) => {
 // @desc    Update user
 // @route   PUT /api/users/:id
 // @access  Private
-const updateUser = asyncHandler(async (req: Request, res: Response) => {
-  const user = await User.findById(req.params.id);
-  if (!user) {
-    res.status(400);
-    throw new Error("User not found");
+const updateUser = asyncHandler(
+  async (req: UserAuthInfoRequest, res: Response) => {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      res.status(404);
+      throw new Error("User not found");
+    }
+    if (req.user.id !== req.params.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
+    const updatedUser = await User.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+    }).select("-password");
+    res.status(200).json(updatedUser);
   }
-  const updatedUser = await User.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-  });
-  res.status(200).json(updatedUser);
-});
+);
 
 // @desc    Delete user
 // @route   DELETE /api/users/:id
 // @access  Private
-const deleteUser = asyncHandler(async (req: Request, res: Response) => {
-  const user = await User.findById(req.params.id);
-  if (!user) {
-    res.status(400);
-    throw new Error("User not found");
+const deleteUser = asyncHandler(
+  async (req: UserAuthInfoRequest, res: Response) => {
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      res.status(404);
+      throw new Error("User not found");
+    }
+    if (req.user.id !== req.params.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
+    await User.findByIdAndDelete(req.params.id);
+    res.status(200).json({ id: req.params.id });
   }
-  await User.findByIdAndDelete(req.params.id);
-  res.status(200).json({ id: req.params.id });
-});
+);
 
 export default {
   getCurrentUser,
